Crate debris_lang

The Debris compiler.

Currently, the compiler is in a very early state. Look at debug_run to see how to compile a script.

debris_std contains the standard library, which is implicitly imported into every script.

The compiler is mostly backend agnostic and only generates a low-level intermediate representation. The crate debris_backends crate contains backend implementations that can convert this ir into files. Right now, the only backend implementation that exists converts the llir code into datapacks. Backends that can create command blocks or even executables might be added in the future.

Functions

debug_run

Compiles the file test.txt into llir

init

main